+++ rust-src/lib.rs
@@ -1,6 +1,6 @@
-mod helpers;
 pub mod aux_functions;
 pub mod external_functions;
+mod helpers;
 pub mod types;
 
 #[macro_use]