+++ src/index.ts
@@ -44,3 +44,4 @@ export {
 } from './credentialDeploymentTransactions';
 export { isAlias, getAlias } from './alias';
 export { deserializeContractState } from './deserialization';
+export * from './blockSummaryHelpers';