+++ src/types.ts
@@ -6,12 +6,11 @@ import { Buffer } from 'buffer/';
 import { ModuleReference } from './types/moduleReference';
 
 /**
- * A reward fraction with a resolution of 1/100000, i.e. the
- * denominator is implicitly 100000, and the interface therefore
- * only contains the numerator value which can be in the interval
- * [1, 100000].
+ * Returns a union of all keys of type T with values matching type V.
  */
-export type RewardFraction = number;
+export type KeysMatching<T, V> = {
+    [K in keyof T]-?: T[K] extends V ? K : never;
+}[keyof T];
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 export interface Versioned<T> {
@@ -257,52 +256,143 @@ export interface FinalizationData {
     finalizers: PartyInfo[];
 }
 
-export interface ExchangeRate {
+export interface Ratio {
     numerator: bigint;
     denominator: bigint;
 }
 
-export interface TransactionFeeDistribution {
-    baker: RewardFraction;
-    gasAccount: RewardFraction;
+export type ExchangeRate = Ratio;
+
+export interface InclusiveRange<N extends number> {
+    min: N;
+    max: N;
 }
 
-export interface MintDistribution {
-    mintPerSlot: number;
-    bakingReward: RewardFraction;
-    finalizationReward: RewardFraction;
+export type DurationSeconds = bigint;
+/** Index of an epoch, or number of epochs. */
+export type Epoch = bigint;
+
+export interface TransactionFeeDistribution {
+    baker: number;
+    gasAccount: number;
 }
+
+export interface MintRate {
+    mantissa: number;
+    exponent: number;
+}
+
+interface MintDistributionCommon {
+    bakingReward: number;
+    finalizationReward: number;
+}
+
+export interface MintDistributionV0 extends MintDistributionCommon {
+    mintPerSlot: number;
+}
+
+export type MintDistributionV1 = MintDistributionCommon;
+
+export type MintDistribution = MintDistributionV0 | MintDistributionV1;
 
 export interface GasRewards {
-    baker: RewardFraction;
-    finalizationProof: RewardFraction;
-    accountCreation: RewardFraction;
-    chainUpdate: RewardFraction;
+    baker: number;
+    finalizationProof: number;
+    accountCreation: number;
+    chainUpdate: number;
 }
 
-export interface RewardParameters {
+interface RewardParametersCommon {
     transactionFeeDistribution: TransactionFeeDistribution;
-    mintDistribution: MintDistribution;
     gASRewards: GasRewards;
 }
 
-export interface ChainParameters {
+/**
+ * Used from protocol version 1-3
+ */
+export interface RewardParametersV0 extends RewardParametersCommon {
+    mintDistribution: MintDistributionV0;
+}
+
+/**
+ * Used from protocol version 4
+ */
+export interface RewardParametersV1 extends RewardParametersCommon {
+    mintDistribution: MintDistributionV1;
+}
+
+export type RewardParameters = RewardParametersV0 | RewardParametersV1;
+
+export interface CooldownParametersV0 {
+    bakerCooldownEpochs: Epoch;
+}
+
+export interface CooldownParametersV1 {
+    poolOwnerCooldown: DurationSeconds;
+    delegatorCooldown: DurationSeconds;
+}
+
+export interface PoolParametersV0 {
+    minimumThresholdForBaking: Amount;
+}
+
+export interface PoolParametersV1 {
+    passiveFinalizationCommission: number;
+    passiveBakingCommission: number;
+    passiveTransactionCommission: number;
+    finalizationCommissionRange: InclusiveRange<number>;
+    bakingCommissionRange: InclusiveRange<number>;
+    transactionCommissionRange: InclusiveRange<number>;
+    minimumEquityCapital: Amount;
+    capitalBound: number;
+    leverageBound: Ratio;
+}
+
+export interface TimeParametersV1 {
+    /**
+     * In epochs
+     */
+    rewardPeriodLength: Epoch;
+    mintPerPayday: number;
+}
+
+interface ChainParametersCommon {
     electionDifficulty: number;
     euroPerEnergy: ExchangeRate;
     microGTUPerEuro: ExchangeRate;
     accountCreationLimit: number;
-    bakerCooldownEpochs: bigint;
-    minimumThresholdForBaking: bigint;
-    rewardParameters: RewardParameters;
     foundationAccountIndex: bigint;
 }
+
+/**
+ * Used from protocol version 1-3
+ */
+export interface ChainParametersV0
+    extends ChainParametersCommon,
+        CooldownParametersV0,
+        PoolParametersV0 {
+    rewardParameters: RewardParametersV0;
+}
+
+/**
+ * Used from protocol version 4
+ */
+export interface ChainParametersV1
+    extends ChainParametersCommon,
+        CooldownParametersV1,
+        PoolParametersV1,
+        TimeParametersV1 {
+    rewardParameters: RewardParametersV1;
+}
+
+export type ChainParameters = ChainParametersV0 | ChainParametersV1;
 
 export interface Authorization {
     threshold: number;
     authorizedKeys: number[];
 }
 
-export interface Authorizations {
+interface AuthorizationsCommon {
     emergency: Authorization;
     microGTUPerEuro: Authorization;
     euroPerEnergy: Authorization;
@@ -311,29 +401,62 @@ export interface Authorizations {
     mintDistribution: Authorization;
     protocol: Authorization;
     paramGASRewards: Authorization;
-    bakerStakeThreshold: Authorization;
+    /**
+     * For protocol version 3 and earlier, this controls the authorization of the bakerStakeThreshold update.
+     */
+    poolParameters: Authorization;
     electionDifficulty: Authorization;
     addAnonymityRevoker: Authorization;
     addIdentityProvider: Authorization;
     keys: VerifyKey[];
 }
 
+/**
+ * Used from protocol version 1-3
+ */
+export type AuthorizationsV0 = AuthorizationsCommon;
+
+/**
+ * Used from protocol version 4
+ */
+export interface AuthorizationsV1 extends AuthorizationsCommon {
+    cooldownParameters: Authorization;
+    timeParameters: Authorization;
+}
+
+export type Authorizations = AuthorizationsV0 | AuthorizationsV1;
+
 export interface KeysWithThreshold {
     keys: VerifyKey[];
     threshold: number;
 }
 
-export interface Keys {
+interface KeysCommon {
     rootKeys: KeysWithThreshold;
     level1Keys: KeysWithThreshold;
-    level2Keys: Authorizations;
 }
+
+/**
+ * Used from protocol version 1-3
+ */
+export interface KeysV0 extends KeysCommon {
+    level2Keys: AuthorizationsV0;
+}
+
+/**
+ * Used from protocol version 4
+ */
+export interface KeysV1 extends KeysCommon {
+    level2Keys: AuthorizationsV1;
+}
+
+export type Keys = KeysV0 | KeysV1;
 
 export interface UpdateQueueQueue {
     effectiveTime: Date;
     // TODO Update the type of update to a generic update transaction when
     // update types have been added.
-    // Information about the actual update.
+    /** Information about the actual update. */
     update: unknown;
 }
 
@@ -342,7 +465,7 @@ export interface UpdateQueue {
     queue: UpdateQueueQueue;
 }
 
-interface UpdateQueues {
+interface UpdateQueuesCommon {
     microGTUPerEuro: UpdateQueue;
     euroPerEnergy: UpdateQueue;
     transactionFeeDistribution: UpdateQueue;
@@ -351,7 +474,6 @@ interface UpdateQueues {
     mintDistribution: UpdateQueue;
     protocol: UpdateQueue;
     gasRewards: UpdateQueue;
-    bakerStakeThreshold: UpdateQueue;
     addAnonymityRevoker: UpdateQueue;
     addIdentityProvider: UpdateQueue;
     rootKeys: UpdateQueue;
@@ -359,17 +481,98 @@ interface UpdateQueues {
     level2Keys: UpdateQueue;
 }
 
-export interface Updates {
-    chainParameters: ChainParameters;
-    keys: Keys;
-    updateQueues: UpdateQueues;
+/**
+ * Used from protocol version 1-3
+ */
+export interface UpdateQueuesV0 extends UpdateQueuesCommon {
+    bakerStakeThreshold: UpdateQueue;
 }
 
-export interface BlockSummary {
+/**
+ * Used from protocol version 4
+ */
+export interface UpdateQueuesV1 extends UpdateQueuesCommon {
+    cooldownParameters: UpdateQueue;
+    timeParameters: UpdateQueue;
+    poolParameters: UpdateQueue;
+}
+
+export type UpdateQueues = UpdateQueuesV0 | UpdateQueuesV1;
+
+interface ProtocolUpdate {
+    message: string;
+    specificationUrl: string;
+    specificationHash: string;
+    specificationAuxiliaryData: string;
+}
+
+interface UpdatesCommon {
+    protocolUpdate: ProtocolUpdate | undefined;
+}
+
+/**
+ * Used from protocol version 1-3
+ */
+export interface UpdatesV0 extends UpdatesCommon {
+    chainParameters: ChainParametersV0;
+    updateQueues: UpdateQueuesV0;
+    keys: KeysV0;
+}
+
+/**
+ * Used from protocol version 4
+ */
+export interface UpdatesV1 extends UpdatesCommon {
+    chainParameters: ChainParametersV1;
+    updateQueues: UpdateQueuesV1;
+    keys: KeysV1;
+}
+
+export type Updates = UpdatesV0 | UpdatesV1;
+
+interface BlockSummaryCommon {
+    protocolVersion?: bigint;
     finalizationData: FinalizationData;
     transactionSummaries: TransactionSummary[];
-    updates: Updates;
 }
+
+/**
+ * Used from protocol version 1-3
+ */
+export interface BlockSummaryV0 extends BlockSummaryCommon {
+    updates: UpdatesV0;
+}
+
+/**
+ * Used from protocol version 4
+ */
+export interface BlockSummaryV1 extends BlockSummaryCommon {
+    updates: UpdatesV1;
+    protocolVersion: bigint;
+}
+
+export type BlockSummary = BlockSummaryV0 | BlockSummaryV1;
+
+interface RewardStatusCommon {
+    protocolVersion?: bigint;
+    totalAmount: Amount;
+    totalEncryptedAmount: Amount;
+    bakingRewardAccount: Amount;
+    finalizationRewardAccount: Amount;
+    gasAccount: Amount;
+}
+
+export type RewardStatusV0 = RewardStatusCommon;
+
+export interface RewardStatusV1 extends RewardStatusCommon {
+    foundationTransactionRewards: Amount;
+    nextPaydayTime: Date;
+    nextPaydayMintRate: MintRate;
+    totalStakedCapital: Amount;
+    protocolVersion: bigint;
+}
+
+export type RewardStatus = RewardStatusV0 | RewardStatusV1;
 
 export interface BlockInfo {
     blockParent: string;
@@ -398,7 +601,13 @@ export interface ConsensusStatus {
     currentEraGenesisBlock: string;
     lastFinalizedBlock: string;
 
+    /**
+     * In milliseconds
+     */
     epochDuration: bigint;
+    /**
+     * In milliseconds
+     */
     slotDuration: bigint;
     bestBlockHeight: bigint;
     lastFinalizedBlockHeight: bigint;
@@ -525,32 +734,229 @@ export interface InitialAccountCredential {
     contents: InitialCredentialDeploymentValues;
 }
 
-export interface BakerReduceStakePendingChange {
-    change: 'ReduceStake';
+export enum StakePendingChangeType {
+    ReduceStake = 'ReduceStake',
+    RemoveStakeV0 = 'RemoveBaker',
+    RemoveStakeV1 = 'RemoveStake',
+}
+
+interface StakePendingChangeV0Common {
+    epoch: bigint;
+}
+
+interface StakePendingChangeV1Common {
+    effectiveTime: Date;
+}
+
+interface ReduceStakePendingChangeCommon {
     newStake: bigint;
-    epoch: bigint;
 }
 
-export interface BakerRemovalPendingChange {
-    change: 'RemoveBaker';
-    epoch: bigint;
+export interface ReduceStakePendingChangeV0
+    extends ReduceStakePendingChangeCommon,
+        StakePendingChangeV0Common {
+    change: StakePendingChangeType.ReduceStake;
 }
 
-export type BakerPendingChange =
-    | BakerReduceStakePendingChange
-    | BakerRemovalPendingChange;
+export interface ReduceStakePendingChangeV1
+    extends ReduceStakePendingChangeCommon,
+        StakePendingChangeV1Common {
+    change: StakePendingChangeType.ReduceStake;
+}
 
-export interface AccountBakerDetails {
+export type ReduceStakePendingChange =
+    | ReduceStakePendingChangeV0
+    | ReduceStakePendingChangeV1;
+
+export interface RemovalPendingChangeV0 extends StakePendingChangeV0Common {
+    change: StakePendingChangeType.RemoveStakeV0;
+}
+
+export interface RemovalPendingChangeV1 extends StakePendingChangeV1Common {
+    change: StakePendingChangeType.RemoveStakeV1;
+}
+
+export type RemovalPendingChange =
+    | RemovalPendingChangeV0
+    | RemovalPendingChangeV1;
+
+export type StakePendingChangeV0 =
+    | ReduceStakePendingChangeV0
+    | RemovalPendingChangeV0;
+
+export type StakePendingChangeV1 =
+    | ReduceStakePendingChangeV1
+    | RemovalPendingChangeV1;
+
+export type StakePendingChange = StakePendingChangeV0 | StakePendingChangeV1;
+
+export enum OpenStatus {
+    OpenForAll = 0,
+    ClosedForNew = 1,
+    ClosedForAll = 2,
+}
+
+/**
+ * How the node translates OpenStatus to JSON.
+ */
+export enum OpenStatusText {
+    OpenForAll = 'openForAll',
+    ClosedForNew = 'closedForNew',
+    ClosedForAll = 'closedForAll',
+}
+
+export type Amount = bigint;
+export type BakerId = bigint;
+
+export interface BakerPoolInfo {
+    openStatus: OpenStatusText;
+    metadataUrl: string;
+    commissionRates: CommissionRates;
+}
+
+export interface CommissionRates {
+    transactionCommission: number;
+    bakingCommission: number;
+    finalizationCommission: number;
+}
+
+export interface CurrentPaydayBakerPoolStatus {
+    blocksBaked: bigint;
+    finalizationLive: boolean;
+    transactionFeesEarned: Amount;
+    effectiveStake: Amount;
+    lotteryPower: number;
+    bakerEquityCapital: Amount;
+    delegatedCapital: Amount;
+}
+
+export enum BakerPoolPendingChangeType {
+    ReduceBakerCapital = 'ReduceBakerCapital',
+    RemovePool = 'RemovePool',
+    NoChange = 'NoChange',
+}
+
+type BakerPoolPendingChangeWrapper<
+    T extends keyof typeof BakerPoolPendingChangeType,
+    S extends Record<string, any>
+> = S & {
+    pendingChangeType: T;
+};
+
+export interface BakerPoolPendingChangeReduceBakerCapitalDetails {
+    bakerEquityCapital: Amount;
+    effectiveTime: Date;
+}
+
+export type BakerPoolPendingChangeReduceBakerCapital =
+    BakerPoolPendingChangeWrapper<
+        BakerPoolPendingChangeType.ReduceBakerCapital,
+        BakerPoolPendingChangeReduceBakerCapitalDetails
+    >;
+
+export interface BakerPoolPendingChangeRemovePoolDetails {
+    effectiveTime: Date;
+}
+
+export type BakerPoolPendingChangeRemovePool = BakerPoolPendingChangeWrapper<
+    BakerPoolPendingChangeType.RemovePool,
+    BakerPoolPendingChangeRemovePoolDetails
+>;
+
+export type BakerPoolPendingChangeNoChange = BakerPoolPendingChangeWrapper<
+    BakerPoolPendingChangeType.NoChange,
+    // eslint-disable-next-line @typescript-eslint/ban-types
+    {}
+>;
+
+export type BakerPoolPendingChange =
+    | BakerPoolPendingChangeReduceBakerCapital
+    | BakerPoolPendingChangeRemovePool
+    | BakerPoolPendingChangeNoChange;
+
+export enum PoolStatusType {
+    BakerPool = 'BakerPool',
+    PassiveDelegation = 'PassiveDelegation',
+}
+
+type PoolStatusWrapper<T extends keyof typeof PoolStatusType, S> = S & {
+    poolType: T;
+};
+
+export interface BakerPoolStatusDetails {
+    bakerId: BakerId;
+    bakerAddress: string;
+    bakerEquityCapital: Amount;
+    delegatedCapital: Amount;
+    delegatedCapitalCap: Amount;
+    poolInfo: BakerPoolInfo;
+    bakerStakePendingChange: BakerPoolPendingChange;
+    currentPaydayStatus?: CurrentPaydayBakerPoolStatus;
+}
+
+export type BakerPoolStatus = PoolStatusWrapper<
+    PoolStatusType.BakerPool,
+    BakerPoolStatusDetails
+>;
+
+export interface PassiveDelegationStatusDetails {
+    delegatedCapital: Amount;
+    commissionRates: CommissionRates;
+    currentPaydayTransactionFeesEarned: Amount;
+    currentPaydayDelegatedCapital: Amount;
+}
+
+export type PassiveDelegationStatus = PoolStatusWrapper<
+    PoolStatusType.PassiveDelegation,
+    PassiveDelegationStatusDetails
+>;
+
+export type PoolStatus = BakerPoolStatus | PassiveDelegationStatus;
+
+export enum DelegationTargetType {
+    PassiveDelegation = 'Passive',
+    Baker = 'Baker',
+}
+
+export interface DelegationTargetPassiveDelegation {
+    delegateType: DelegationTargetType.PassiveDelegation;
+}
+
+export interface DelegationTargetBaker {
+    delegateType: DelegationTargetType.Baker;
+    bakerId: BakerId;
+}
+
+export type DelegationTarget =
+    | DelegationTargetPassiveDelegation
+    | DelegationTargetBaker;
+
+interface AccountBakerDetailsCommon {
     restakeEarnings: boolean;
-    bakerId: bigint;
+    bakerId: BakerId;
     bakerAggregationVerifyKey: string;
     bakerElectionVerifyKey: string;
     bakerSignatureVerifyKey: string;
     stakedAmount: bigint;
-    pendingChange?: BakerPendingChange;
+    pendingChange?: StakePendingChange;
 }
 
-export interface AccountInfo {
+export type AccountBakerDetailsV0 = AccountBakerDetailsCommon;
+
+export interface AccountBakerDetailsV1 extends AccountBakerDetailsCommon {
+    bakerPoolInfo: BakerPoolInfo;
+}
+
+export type AccountBakerDetails = AccountBakerDetailsV0 | AccountBakerDetailsV1;
+
+export interface AccountDelegationDetails {
+    restakeEarnings: boolean;
+    stakedAmount: bigint;
+    delegationTarget: DelegationTarget;
+    pendingChange?: StakePendingChangeV1;
+}
+
+interface AccountInfoCommon {
     accountNonce: bigint;
     accountAmount: bigint;
     accountIndex: bigint;
@@ -566,9 +972,30 @@ export interface AccountInfo {
         number,
         Versioned<InitialAccountCredential | NormalAccountCredential>
     >;
-
-    accountBaker?: AccountBakerDetails;
 }
+
+export type AccountInfoSimple = AccountInfoCommon;
+
+export interface AccountInfoBakerV0 extends AccountInfoCommon {
+    accountBaker: AccountBakerDetailsV0;
+}
+
+/** Protocol version 4 and later. */
+export interface AccountInfoBakerV1 extends AccountInfoCommon {
+    accountBaker: AccountBakerDetailsV1;
+}
+
+export type AccountInfoBaker = AccountInfoBakerV0 | AccountInfoBakerV1;
+
+/** Protocol version 4 and later. */
+export interface AccountInfoDelegator extends AccountInfoCommon {
+    accountDelegation: AccountDelegationDetails;
+}
+
+export type AccountInfo =
+    | AccountInfoSimple
+    | AccountInfoBaker
+    | AccountInfoDelegator;
 
 export interface Description {
     name: string;
@@ -620,6 +1047,7 @@ export enum AccountTransactionType {
     SimpleTransferWithMemo = 22,
     EncryptedTransferWithMemo = 23,
     TransferWithScheduleWithMemo = 24,
+    ConfigureDelegation = 26,
 }
 
 export interface DeployModulePayload {
@@ -724,6 +1152,15 @@ export interface UpdateCredentialsPayload {
     currentNumberOfCredentials: bigint;
 }
 
+export interface ConfigureDelegationPayload {
+    /* stake to delegate. if set to 0, this removes the account as a delegator */
+    stake?: GtuAmount;
+    /* should earnings from delegation be added to staked amount  */
+    restakeEarnings?: boolean;
+    /* determines if the account should use passive delegation, or which specific baker to delegate to  */
+    delegationTarget?: DelegationTarget;
+}
+
 export type AccountTransactionPayload =
     | SimpleTransferPayload
     | SimpleTransferWithMemoPayload
@@ -731,7 +1168,8 @@ export type AccountTransactionPayload =
     | DeployModulePayload
     | InitContractPayload
     | UpdateContractPayload
-    | UpdateCredentialsPayload;
+    | UpdateCredentialsPayload
+    | ConfigureDelegationPayload;
 
 export interface AccountTransaction {
     type: AccountTransactionType;