+++ test/accountTransaction.test.ts
@@ -12,6 +12,8 @@ import {
     SimpleTransferWithMemoPayload,
     UpdateCredentialsPayload,
     VerifyKey,
+    ConfigureDelegationPayload,
+    DelegationTargetType,
 } from '../src/types';
 import * as ed from 'noble-ed25519';
 import {
@@ -279,6 +281,58 @@ test('send registerData signed with wrong private key is accepted', async () => 
 
     const result = await client.sendAccountTransaction(
         registerDataTransaction,
+        signatures
+    );
+    expect(result).toBeTruthy();
+});
+
+test('send configureDelegation signed with wrong private key is accepted', async () => {
+    const senderAccountAddress =
+        '4ZJBYQbVp3zVZyjCXfZAAYBVkJMyVj8UKUNj9ox5YqTCBdBq2M';
+
+    const nextAccountNonce = await client.getNextAccountNonce(
+        new AccountAddress(senderAccountAddress)
+    );
+    if (!nextAccountNonce) {
+        throw new Error('Nonce not found!');
+    }
+    const header: AccountTransactionHeader = {
+        expiry: new TransactionExpiry(new Date(Date.now() + 3600000)),
+        nonce: nextAccountNonce.nonce,
+        sender: new AccountAddress(senderAccountAddress),
+    };
+    const configureDelegationPayload: ConfigureDelegationPayload = {
+        stake: new GtuAmount(1000000000n),
+        delegationTarget: {
+            delegateType: DelegationTargetType.Baker,
+            bakerId: 0n,
+        },
+        restakeEarnings: true,
+    };
+
+    const configureDelegationTransaction: AccountTransaction = {
+        header,
+        payload: configureDelegationPayload,
+        type: AccountTransactionType.ConfigureDelegation,
+    };
+
+    const wrongPrivateKey =
+        'ce432f6cca0d47caec1f45739331dc354b6d749fdb8ab7c2b7f6cb24db39ca0c';
+
+    const hashToSign = getAccountTransactionSignDigest(
+        configureDelegationTransaction
+    );
+    const signature = Buffer.from(
+        await ed.sign(hashToSign, wrongPrivateKey)
+    ).toString('hex');
+    const signatures: AccountTransactionSignature = {
+        0: {
+            0: signature,
+        },
+    };
+
+    const result = await client.sendAccountTransaction(
+        configureDelegationTransaction,
         signatures
     );
     expect(result).toBeTruthy();