+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+
+## 0.8.0 2022-05-??
+
+### Added
+
+- Support for getting baker list from node.
+- Support for getting status of a Baker Pool/Passive delegation (required node to have protocol version 4 or later).
+- Support for getting reward status of chain at specific block.
+- Helper functions for determining the version of `BlockSummary` and nested types. 
+- Helper functions for determining the version of `AccountInfo` variants. 
+- Support for the new "configure delegation" transaction type.
+
+### Changed
+
+- Updated `BlockSummary` type to include new version, effective from protocol version 4.
+- Updated `AccountInfo` type to include new fields related to delegation introduced with protocol version 4.
+
 ## 0.7.3 2022-05-05
 
 ### Added